ai_job_agent/requirements.txt



openai>=1.2.4
python-dotenv
